requirements.txt
Flask
psutil
flask-cors
google-generativeai